% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRank-package.R
\docType{package}
\name{BayesRank-package}
\alias{BayesRank}
\alias{BayesRank-package}
\title{BayesRank: Bayesian Rank-Based Isotonic Regression for Caenorhabditis elegans Assays}
\description{
Implements a Bayesian isotonic rank-likelihood model for ordinal regression
under monotonicity constraints, developed to analyze Caenorhabditis elegans neuron
damage scores from a developmental rotenone exposure assay from
\href{https://academic.oup.com/toxsci/article/209/3/kfag011/8475413}{Bergemann et al. (2026) "\emph{Progeny effects of rotenone exposure depend on parental toxicity}"}.
The package reproduces the analysis from \href{https://arxiv.org/abs/PLACEHOLDER}{Presman, Anceschi, Huayta, Meyer, and Herring, (2026+) "\emph{Order-Restricted Bayesian Ordinal Regression for the Modeling of Neuron Degeneration in Caenorhabditis elegans}"}.
}
\details{
Relative to existing rank-likelihood implementations, it additionally supports
random effects, parameter constraints, collapsed sampling, and parameter expansion.
While the bundled data, design matrices, and tutorial are specific to
the C. elegans assay design described above, the underlying isotonic
rank-likelihood methodology readily generalizes to a broad range of rank-based
regression settings.
}
\author{
\strong{Maintainer}: Niccolo Anceschi \email{niccolo.anceschi@duke.edu}

Authors:
\itemize{
  \item Niccolo Anceschi \email{niccolo.anceschi@duke.edu}
}

Other contributors:
\itemize{
  \item Rick Presman \email{rick.presman@duke.edu} [contributor]
}

}
\keyword{internal}
