% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeSRank.R
\name{BayeSRank}
\alias{BayeSRank}
\title{BayeSRank: Bias-Aware Rank Aggregation}
\usage{
BayeSRank(x)
}
\arguments{
\item{x}{This is an}
}
\description{
This is the bespoke BayeSRank algorithm introduced in Ruan et al (XXXX) paper.
This implementation does not contain the two-step procedure. If you love
two-stepping, please use XXX.
}
\seealso{
The two-step function.
}
