% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot_latent_scores}
\alias{plot_latent_scores}
\title{Plot category-wise posterior distributions of latent scores}
\usage{
plot_latent_scores(
  draws_mcmc,
  show = T,
  save = F,
  filename = "hist_Z_scores",
  dir = "~/Desktop"
)
}
\arguments{
\item{draws_mcmc}{Postprocessed MCMC output from \code{postprocess_BayesRank}}

\item{show}{Display the plot to the active graphics device? Default TRUE.}

\item{save}{Save the plot to a PDF file? Default FALSE.}

\item{filename}{Output file name (without extension). Default 'hist_Z_scores'.}

\item{dir}{Output directory. Default '~/Desktop'.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plots overlapping histograms of the latent utilities, groupe and colored by
observed ordinal category, to visualize separation between adjacent groups.
}
