% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gibbs.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Check and fill defaults for Gibbs sampler hyperparameters}
\usage{
set_hyperparameters(hyper_params = NULL)
}
\arguments{
\item{hyper_params}{List of hyperparameters, with elements:
\describe{
\item{w0}{Prior SD for normla prior on the group-specific intercepts. Default 1.}
\item{pi0}{Prior spike probability for treatment coefficients. Default 0.5.}
\item{mu0}{Prior mean for the gaussian slab component of treatment coefficients. Default 0.}
\item{lambda0}{Prior SD for the gaussian slab component of treatment coefficients. Default 1.}
\item{r_shape}{Shape hyperparameter for the inverse-gamma prior on the random-effect variance (rho2). Default 2.}
\item{r_rate}{Rate hyperparameter for the inverse-gamma prior on the random-effect variance (rho2). Default 2.}
\item{iter_MH}{Inner sweeps for multivariate truncated normal Gibbs sampling. Default 10.}
\item{rho2_GG}{Initial/center value for Griddy-Gibbs on rho2. Default 2.}
\item{rho2_max}{Upper bound for Griddy-Gibbs on rho2. Default 5.}
\item{n_grid_rho2}{Number of grid points for Griddy-Gibbs on rho2. Default 100.}
}
May be partially specified or NULL; missing elements are filled with defaults.}
}
\value{
A complete list of hyperparameters.
}
\description{
Validates a list of hyperparameters for \code{gibbs_sampler}, filling in
default values for any missing elements.
}
